Fixed hit packet handler to handle the correct key in 0.75.
